pyproject.toml
[tool.poetry]
name = "odb-app"
version = "0.1.0"
description = ""
authors = ["fooqoo <fooqoofooqoo56@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
customtkinter = "^5.2.2"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
